Agregar nuevas páginas y componentes al frontend, incluyendo secciones de servicios, proyectos privados, obras civiles y profesionales. Se implementan mejoras en la gestión de imágenes y se optimizan los formularios de creación y edición de proyectos. Se añaden animaciones y efectos visuales para mejorar la experiencia del usuario. Además, se actualizan dependencias y se corrigen errores menores en el código.

+++ constructora-temuco-backend/config/app.php
@@ -136,7 +136,7 @@ return [
 
     'max_image_size' => env('MAX_IMAGE_SIZE', 10240),
     'allowed_image_types' => env('ALLOWED_IMAGE_TYPES', 'jpg,jpeg,png,webp'),
-    'image_quality' => env('IMAGE_QUALITY', 85),
+    'image_quality' => (int) env('IMAGE_QUALITY', 85),
     'thumbnail_size' => env('THUMBNAIL_SIZE', 300),
 
 ];
